Changed entity mapping. Configured services work

+++ EntityMapping/EntityMapping/Entities/Fields.cs
@@ -2,11 +2,11 @@
 {
     public class Fields
     {
-        public int Id { get; set; }
+        public byte Id { get; set; }
 
         public string Name { get; set; }
 
-        public int ListId { get; set; }
+        public short ListId { get; set; }
 
         public virtual Lists Lists { get; set; }
     }
